backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
firebase-admin>=6.1.0
pydantic>=2.0.0
httpx>=0.24.0
pillow>=9.5.0
pyrebase4>=4.6.0
requests>=2.28.2
python-jose>=3.3.0
passlib>=1.7.4
tenacity>=8.2.2